update almost all page

+++ books/delete.php
@@ -1,16 +1,26 @@
-<?php 
+<?php
 
 require __DIR__ . '/../config/database.php';
+
+session_start();
 
 // $id = $_GET['id'];
 $id = $_POST['id'];
 
-// var_dump($id);
+$sql_check = "SELECT COUNT(*) AS 'Jumlah' FROM borrowings WHERE book_id = $id AND status = 'Dipinjam'";
+$result = $conn->query($sql_check)->fetch_assoc();
 
-$sql = "DELETE FROM books WHERE id=$id";
-$conn->query($sql);
+$jumlah = $result['Jumlah'];
+
+// var_dump($result);
+
+if ($jumlah == 0) {
+    $sql = "DELETE FROM books WHERE id=$id";
+    $conn->query($sql);
+    $_SESSION['buku-dihapus'] = "Buku berhasil dihapus";
+} else {
+    $_SESSION['buku-gagal-dihapus'] = "Buku gagal dihapus";
+}
 
 header("Location: /../books/index.php");
 exit();
-
-?>
